EDSC-3779: Cleanup
