+++ static/src/js/actions/index.js
@@ -10,13 +10,9 @@ import {
 } from './admin/retrievals'
 
 import {
-  // adminViewMetricsRetrieval,
   fetchAdminMetricsRetrievals,
-  // fetchAdminMetricsRetrieval,
-  // requeueMetricsOrder,
   updateAdminMetricsRetrievalsStartDate,
   updateAdminMetricsRetrievalsEndDate
-  // updateAdminMetricsRetrievalsPageNum
 } from './admin/retrievalMetrics'
 
 import {
@@ -320,13 +316,9 @@ const actions = {
   updateAdminRetrievalsPageNum,
   updateAdminRetrievalsSortKey,
   // todo new actions
-  // adminViewMetricsRetrieval,
   fetchAdminMetricsRetrievals,
-  // fetchAdminMetricsRetrieval,
-  // requeueMetricsOrder,
   updateAdminMetricsRetrievalsStartDate,
   updateAdminMetricsRetrievalsEndDate,
-  // updateAdminMetricsRetrievalsPageNum,
   // todo new actions
   updateAdvancedSearch,
   updateAuthToken,
